requirements.txt
pdfkit==1.0.0
markdown2==2.5.0
black
isort
pydantic==2.9.2
pydantic-settings==2.5.2
pandas==2.2.2
openai==1.46.0
langchain-experimental==0.3.0
openpyxl==3.1.5
matplotlib==3.9.2
plotly-express==0.4.1
kaleido
langgraph==0.2.23
langchain==0.3.0
langchain-openai==0.2.0
beautifulsoup4==4.12.3
langfuse==2.51.2
xlrd==2.0.1
weasyprint==62.3
pyppeteer==2.0.0
pytest-playwright==0.5.2
pypandoc-binary==1.14
python-docx==1.1.2